Clarify README
There was written something about the 'modification' of files. Make that
refer to the ctime.

Why the ctime? On my computer I have got a directory named 5 for files
that shall be deleted after five days of non-modification. However, if I
have a file lying around somewhere else and decide that it should be
deleted in five days and therefore put it in 5, it may well be that
mtime says the file hasn't been modified for 5 days and it is deleted
immediately.

It probably should be an option to the user which time to use. But not
now.